chore(ci): bump the pinned Rust toolchain to 1.97.1
The pin sat at 1.93.1 (2026-02-11), four releases behind stable 1.97.1
(2026-07-14). Bump it and move the `dtolnay/rust-toolchain` tag in
`ci.yml` with it, as the comment on that step requires.

The `@stable` steps in `release.yml`, `nightly-verify.yml`, and
`pipeline-parallel-ci.yml` were never building at a different version:
the action runs `rustup default` and never exports `RUSTUP_TOOLCHAIN`,
which `rust-toolchain.toml` overrides per directory, so every cargo
invocation in the tree already resolved to the pin.

`cargo fmt` produces no diff at 1.97.1, so the bump reformats nothing.
Six new clippy lints fire under `-D warnings` and are fixed here, all
mechanical and behavior-preserving:

  question_mark              memory_estimate.rs, sanitize.rs
  collapsible_match          chat_request.rs
  for_kv_map                 request_router_tests.rs
  unnecessary_cast           dbrx_tests.rs
  unneeded_wildcard_pattern  pipeline_remote_real_models.rs

+++ src/models/dbrx_tests.rs
@@ -314,9 +314,7 @@ fn clip_qkv_actually_clamps_the_projection() {
     assert_eq!(unclipped.clip_qkv, None);
 
     // Large inputs push the fused projection well past +/-8.
-    let big: Vec<f32> = (0..(args.d_model as usize))
-        .map(|i| 40.0 + i as f32)
-        .collect();
+    let big: Vec<f32> = (0..args.d_model).map(|i| 40.0 + i as f32).collect();
     let x = mlxcel_core::from_slice_f32(&big, &[1, 1, args.d_model as i32]);
 
     let mut cache_a = KVCache::new();
